requirements.txt
bottle
beautifulsoup4
requests
python-bitcoinrpc
